RestrictionsProvider: map long values to int
Restriction entries don't have a Long type, only Integer [1].
Some settings values are of type long, such as "default_sync_interval".
Since we want to provide those via managed configurations, map long
values to int. This needs to be done carefully and only for values which
don't exceed int. In case of "default_sync_interval" it should be
unproblematic, as INT_MAX translations to an interval of 68 years and we
shouldn't exceed that.

[1] https://developer.android.com/reference/android/content/RestrictionEntry#TYPE_INTEGER

+++ app/src/main/java/com/messageconcept/peoplesyncclient/settings/RestrictionsProvider.kt
@@ -70,7 +70,7 @@ class RestrictionsProvider(
             getValue(key) { restrictions -> restrictions.getInt(key) }
 
     override fun getLong(key: String) =
-            getValue(key) { restrictions -> restrictions.getLong(key) }
+            getValue(key) { restrictions -> restrictions.getInt(key).toLong() }
 
     override fun getString(key: String) =
             getValue(key) { restrictions -> restrictions.getString(key) }
